# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS API Documentation",
    "version": "1.0.0",
    "description": "API documentation for EduNexa LMS"
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "tags": [
    { "name": "Authentication APIs", "description": "User registration and login for Students and Instructors" },
    { "name": "Admin Authentication APIs", "description": "Admin-only registration and login endpoints" },
    { "name": "User APIs", "description": "User profile management" },
    { "name": "Course APIs", "description": "Course management - Public access for viewing, Admin only for create/update/delete" },
    { "name": "Enrollment APIs", "description": "Student course enrollment and my courses" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "POST /auth/register - User Registration",
        "description": "Register a new student or regular user account. Default role is 'student'.",
        "tags": ["Authentication APIs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "User already exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "POST /auth/login - User Login",
        "description": "Login for students and regular users. Returns JWT token and user info.",
        "tags": ["Authentication APIs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/admin/register": {
      "post": {
        "summary": "POST /auth/admin/register - Admin Registration",
        "description": "Register a new administrator account. Automatically sets role to 'admin'.",
        "tags": ["Admin Authentication APIs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Admin User" },
                  "email": { "type": "string", "example": "admin@edunexa.com" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Admin registered successfully" },
          "400": { "description": "User already exists" }
        }
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "summary": "POST /auth/admin/login - Admin Login",
        "description": "Admin-only login. Validates that user has admin role before allowing access.",
        "tags": ["Admin Authentication APIs"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "admin@edunexa.com" },
                  "password": { "type": "string", "example": "admin123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Admin login successful" },
          "401": { "description": "Invalid credentials or not an admin" }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "GET /auth/profile - Get User Profile",
        "tags": ["User APIs"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile data" },
          "401": { "description": "Not authorized" }
        }
      },
      "put": {
        "summary": "PUT /auth/profile - Update User Profile",
        "tags": ["User APIs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated successfully" },
          "401": { "description": "Not authorized" }
        }
      }
    },
    "/api/courses": {
      "get": {
        "summary": "GET /courses - Get All Courses",
        "tags": ["Course APIs"],
        "responses": {
          "200": { "description": "List of courses" }
        }
      },
      "post": {
        "summary": "POST /courses - Create Course (Admin only)",
        "tags": ["Course APIs"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "category"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "category": { "type": "string" },
                  "level": { "type": "string", "enum": ["beginner", "intermediate", "advanced"] },
                  "price": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Course created" },
          "403": { "description": "Not authorized" }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "summary": "GET /courses/:id - Get Course by ID",
        "tags": ["Course APIs"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Course details" },
          "404": { "description": "Course not found" }
        }
      },
      "put": {
        "summary": "PUT /courses/:id - Update Course (Admin only)",
        "tags": ["Course APIs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "category": { "type": "string" },
                  "level": { "type": "string" },
                  "price": { "type": "number" },
                  "isPublished": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Course updated" },
          "403": { "description": "Not authorized" },
          "404": { "description": "Course not found" }
        }
      },
      "delete": {
        "summary": "DELETE /courses/:id - Delete Course (Admin only)",
        "tags": ["Course APIs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Course deleted" },
          "403": { "description": "Not authorized" },
          "404": { "description": "Course not found" }
        }
      }
    },
    "/api/enroll/{courseId}": {
      "post": {
        "summary": "POST /enroll/:courseId - Enroll in Course",
        "tags": ["Enrollment APIs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "courseId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "201": { "description": "Successfully enrolled" },
          "400": { "description": "Already enrolled" },
          "401": { "description": "Not authorized" },
          "404": { "description": "Course not found" }
        }
      }
    },
    "/api/my-courses": {
      "get": {
        "summary": "GET /my-courses - Get My Enrolled Courses",
        "tags": ["Enrollment APIs"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of enrolled courses" },
          "401": { "description": "Not authorized" }
        }
      }
    }
  }
}